keep-bridge/requirements.txt
Flask==3.1.0
gkeepapi==0.16.0
gpsoauth==2.0.0
python-dotenv==1.0.1
bleak>=0.22,<1.0
